Manshree Constructions Website
The main objective of the project is to showcase the current projects being constructed under Manshree developers and to highlight the benefits, amenities and property rates to the customers. Through the website the customers can have a quick visit over the property frames, inner design, location, built-area and amenities provided as a whole. The exact location of the current project is pinned into the website via Google Maps and on top of that it also displays the graphic location of project which covers the near-by areas as well.

The website was build using HTML 5, CSS, Javascript, bootstrap along with  php as database